requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==4.4.2
distro==1.9.0
exceptiongroup==1.2.0
ffmpeg==1.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.0
imageio-ffmpeg==0.4.9
moviepy==1.0.3
mutagen==1.47.0
numpy==1.26.4
openai==1.17.1
pexels-api==1.0.1
pillow==10.3.0
proglog==0.1.10
pydantic==2.7.0
pydantic_core==2.18.1
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
